Seperated out the main and app logic (#502)
