+++ d2app/capture_state.go
@@ -1,0 +1,9 @@
+package d2app
+
+type captureState int
+
+const (
+	captureStateNone captureState = iota
+	captureStateFrame
+	captureStateGif
+)
